TestVoltProcedure compiles

+++ tests/frontend/org/voltdb/TestVoltProcedure.java
@@ -398,7 +398,8 @@ public class TestVoltProcedure extends TestCase {
         assertNotNull(agent.m_selector);
         assertNotNull(agent.m_source);
         assertEquals(agent.m_selector, SysProcSelector.PROCEDURE);
-        assertEquals(agent.m_catalogId, Integer.parseInt(site.m_context.cluster.getSites().get(Integer.toString(site.getSiteId())).getTypeName()));
+        assertEquals(agent.m_catalogId,
+                     Integer.parseInt(site.m_context.cluster.getSites().get(Long.toString(site.getSiteId())).getTypeName()));
         Object statsRow[][] = agent.m_source.getStatsRows(false, 0L);
         assertNotNull(statsRow);
         assertEquals( 0, statsRow.length);
